Tooltips
Adds a primitive, wonky tooltip function that really sucks.
Descriptions are bad and unhelpful.
I'd rather find a way to incorporate information into the bottom right
HUD thing, but I'm not a GUI guy.
Only works for weapons, can work on orders if the terrible 'TEXT THAT
JUST FLOATS IN THE MIDDLE OF NOWHERE" thing is used and doesn't suck too
much alongside the rest of the functions.